wip(weather): rename and move Weather compoent; adjust imports; make `settings` optional on widget settings object (#28)
