+++ src/lib/widgets/Weather/WeatherIcon.svelte
