+++ src/lib/widgets/Weather/weather.css
